added IntervalDemo that shows how to use the scheduler, and how pause/resume works. Scheduler uses float instead of double (faster) Scheduler/CocosNode supports "[schedule .... ... interval:(float)]"
don't consume cpu while paused seems that it is not working... invistigate.

+++ cocos2d/Particle.h
@@ -150,7 +150,7 @@ typedef struct sParticle
 //! Add a particle to the emitter
 -(BOOL) addParticle;
 //! Update all particles
--(void) step:(double)dt;
+-(void) step:(float)dt;
 //! Initializes a particle
 -(void) initParticle: (Particle*) particle;
 //! draw all the particles
